Vehicle delete - check if there are related bookings

+++ src/Controller/VehicleController.php
@@ -2,13 +2,18 @@
 
 namespace App\Controller;
 
+use App\Entity\Booking;
 use App\Entity\Vehicle;
 use App\Form\VehicleType;
 use App\Repository\VehicleRepository;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\OptimisticLockException;
+use Doctrine\ORM\ORMException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Service\Attribute\Required;
 
 #[Route('/vehicle')]
 class VehicleController extends AbstractController
@@ -19,6 +24,9 @@ class VehicleController extends AbstractController
     public const ROUTE_SHOW = 'vehicle_show';
     public const ROUTE_DELETE = 'vehicle_delete';
 
+    #[Required]
+    public EntityManagerInterface $entityManager;
+
     #[Route('/', name: self::ROUTE_INDEX, methods: ['GET'])]
     public function index(VehicleRepository $vehicleRepository): Response
     {
@@ -27,6 +35,10 @@ class VehicleController extends AbstractController
         ]);
     }
 
+    /**
+     * @throws OptimisticLockException
+     * @throws ORMException
+     */
     #[Route('/new', name: self::ROUTE_NEW, methods: ['GET', 'POST'])]
     public function new(Request $request, VehicleRepository $vehicleRepository): Response
     {
@@ -36,7 +48,7 @@ class VehicleController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $vehicleRepository->add($vehicle);
-            return $this->redirectToRoute('app_vehicle_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute(self::ROUTE_INDEX, [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('vehicle/new.html.twig', [
@@ -53,6 +65,10 @@ class VehicleController extends AbstractController
         ]);
     }
 
+    /**
+     * @throws OptimisticLockException
+     * @throws ORMException
+     */
     #[Route('/{id}/edit', name: self::ROUTE_EDIT, methods: ['GET', 'POST'])]
     public function edit(Request $request, Vehicle $vehicle, VehicleRepository $vehicleRepository): Response
     {
@@ -61,7 +77,7 @@ class VehicleController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $vehicleRepository->add($vehicle);
-            return $this->redirectToRoute('app_vehicle_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute(self::ROUTE_INDEX, [], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('vehicle/edit.html.twig', [
@@ -70,13 +86,31 @@ class VehicleController extends AbstractController
         ]);
     }
 
+    /**
+     * @throws OptimisticLockException
+     * @throws ORMException
+     */
     #[Route('/{id}/delete', name: self::ROUTE_DELETE, methods: ['POST'])]
     public function delete(Request $request, Vehicle $vehicle, VehicleRepository $vehicleRepository): Response
     {
+        $relatedBookings = $this->entityManager->getRepository(Booking::class)->findBy(['vehicle' => $vehicle->getId()]);
+
+        if ($relatedBookings) {
+            $bookingIds = [];
+
+            foreach ($relatedBookings as $booking) {
+                $bookingIds[] = $booking->getId();
+            }
+
+            $this->addFlash('error', 'This vehicle is used on booking(s) with id: ' . implode(', ', $bookingIds));
+
+            return $this->redirectToRoute(self::ROUTE_SHOW, ['id' => $vehicle->getId()], Response::HTTP_SEE_OTHER);
+        }
+
         if ($this->isCsrfTokenValid('delete'.$vehicle->getId(), $request->request->get('_token'))) {
             $vehicleRepository->remove($vehicle);
         }
 
-        return $this->redirectToRoute('app_vehicle_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute(self::ROUTE_INDEX, [], Response::HTTP_SEE_OTHER);
     }
 }
